# dict cache optimization